refactor(test): split one test file into multiple, to better organize
